requirements.txt
streamlit
fpdf
python-dotenv
requests
